Mediquery_sonar_api/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
python-dotenv==1.0.0
pydantic==2.5.3
astroid>=2.14.2
pylint>=2.17.0 